# openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "pi-agent-server host API",
    "version": "1.0.0",
    "description": "Fixed host API contract v1. Dynamic plugin routes are not part of this document.\nRequest validation semantics: Request bodies use Fastify's AJV baseline: declared types are coerced (`coerceTypes: 'array'`,\nso numbers, booleans, and single-element arrays are accepted where a string is declared) and\nundeclared fields are silently stripped (`removeAdditional: true` combined with\n`additionalProperties: false`) instead of returning 400. A 400 therefore reports a missing\nrequired field or a value that cannot be coerced/constrained, never a wrong scalar type or an\nunknown field."
  },
...
  "security": [
    {
      "IpAdmission": []
    },
    {
      "IpAdmission": [],
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/access": {
      "get": {
        "operationId": "getAccess",
        "security": [
          {
            "IpAdmission": []
          },
          {
            "IpAdmission": [],
            "BearerAuth": []
          }
        ],
        "x-pi-rbac": {
          "permission": "access:read",
          "roles": [
            "admin",
            "user",
            "viewer"
          ]
        },
        "responses": {
          "200": {
            "description": "Access projection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Access"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token required for this admitted IP profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "IP admission or RBAC denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "head": {
        "operationId": "getAccessHead",
        "security": [
          {
            "IpAdmission": []
          },
          {
            "IpAdmission": [],
            "BearerAuth": []
          }
        ],
        "x-pi-rbac": {
          "permission": "access:read",
          "roles": [
            "admin",
            "user",
            "viewer"
          ]
        },
        "responses": {
          "200": {
            "description": "Access projection; the GET response body is suppressed.",
            "headers": {
              "Content-Type": {
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "required": true,
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token required for this admitted IP profile.",
            "headers": {
              "Content-Type": {
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "required": true,
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              },
              "WWW-Authenticate": {
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "Bearer"
                }
              }
            }
          },
          "403": {
            "description": "IP admission or RBAC denied.",
            "headers": {
              "Content-Type": {
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "required": true,
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sessions": {
      "get": {
        "operationId": "listSessions",
        "security": [
          {
            "IpAdmission": []
          },
          {
            "IpAdmission": [],
            "BearerAuth": []
          }
        ],
        "x-pi-rbac": {
          "permission": "sessions:list",
          "roles": [
            "admin",
            "user",
            "viewer"
          ]
        },
        "responses": {
          "200": {
            "description": "Sessions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Bearer token required for this admitted IP profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "IP admission or RBAC denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "head": {
        "operationId": "listSessionsHead",
        "security": [
          {
            "IpAdmission": []
          },
          {
            "IpAdmission": [],
            "BearerAuth": []
          }
        ],
        "x-pi-rbac": {
          "permission": "sessions:list",
          "roles": [
            "admin",
            "user",
            "viewer"
          ]
        },
        "responses": {
          "200": {
            "description": "Sessions; the GET response body is suppressed.",
            "headers": {
              "Content-Type": {
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "required": true,
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token required for this admitted IP profile.",
            "headers": {
              "Content-Type": {
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "required": true,
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              },
              "WWW-Authenticate": {
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "Bearer"
                }
              }
            }
          },
          "403": {
            "description": "IP admission or RBAC denied.",
            "headers": {
              "Content-Type": {
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "required": true,
                "schema": {
                  "type": "string",
                  "pattern": "^[0-9]+$"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "createSession",
        "security": [
          {
            "IpAdmission": []
          },
          {
            "IpAdmission": [],
            "BearerAuth": []
          }
        ],
        "x-pi-rbac": {
          "permission": "sessions:create",
          "roles": [
            "admin",
            "user"
          ]
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Invalid model or thinking level.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token required for this admitted IP profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "IP admission or RBAC denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Model availability check failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSession"
              }
            }
          }
        },
        "x-pi-request-validation": {
          "coerceTypes": true,
          "removeAdditional": "silent-strip"
        },
        "description": "Request bodies use Fastify's AJV baseline: declared types are coerced (`coerceTypes: 'array'`,\nso numbers, booleans, and single-element arrays are accepted where a string is declared) and\nundeclared fields are silently stripped (`removeAdditional: true` combined with\n`additionalProperties: false`) instead of returning 400. A 400 therefore reports a missing\nrequired field or a value that cannot be coerced/constrained, never a wrong scalar type or an\nunknown field."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "IpAdmission": {
        "type": "http",
        "scheme": "ip-admission",
        "description": "Mandatory direct-peer CIDR/IP admission. This is evaluated by the host, not supplied by a client.",
        "x-pi-ip-admission": true
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Required only when the admitted IP profile has tokenRequired=true."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "statusCode",
          "error",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "description": "Fastify validation error code when applicable."
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Access": {
        "type": "object",
        "required": [
          "canRead",
          "canWrite"
        ],
        "properties": {
          "canRead": {
            "type": "boolean"
          },
          "canWrite": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Session": {
        "type": "object",
        "required": [
          "id",
          "ownerKey",
          "projectId",
          "title",
          "createdAt",
          "updatedAt",
          "modelProvider",
          "modelId",
          "thinkingLevel",
          "systemPrompt",
          "capabilityVersions"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerKey": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "createdAt": {
            "type": "number"
          },
          "updatedAt": {
            "type": "number"
          },
          "modelProvider": {
            "type": [
              "string",
              "null"
            ]
          },
          "modelId": {
            "type": [
              "string",
              "null"
            ]
          },
          "thinkingLevel": {
            "type": [
              "string",
              "null"
            ]
          },
          "systemPrompt": {
            "type": [
              "string",
              "null"
            ]
          },
          "capabilityVersions": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "CreateSession": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "modelProvider": {
            "type": "string",
            "minLength": 1
          },
          "modelId": {
            "type": "string",
            "minLength": 1
          },
          "thinkingLevel": {
            "type": "string",
            "enum": [
              "off",
              "minimal",
              "low",
              "medium",
              "high",
              "xhigh",
              "max"
            ]
          }
        },
        "dependentRequired": {
          "modelProvider": [
            "modelId"
          ],
          "modelId": [
            "modelProvider"
          ]
        },
        "additionalProperties": false
      },
...
    }
  }
}